+++ webpack.prod.config.js
@@ -46,6 +46,10 @@ module.exports = {
 	},
 	plugins: [
 		new webpack.NoErrorsPlugin(),
+		new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify('production'),
+      '__DEV__': false
+    }),
 		new webpack.optimize.UglifyJsPlugin({
 			// js压缩: 已经压缩过的文件是否还会被压缩？
 			compress: {