+++ webpackVersion/scripts/links/web-source.js
@@ -1,0 +1,45 @@
+module.exports = {
+  title: 'web source',
+  contents: [
+    {
+      title: '知乎',
+      date: '2016-08-13',
+      url: 'https://www.zhihu.com/people/zjdgx-george/activities'
+    },
+    {
+      title: '凹凸实验室',
+      date: '2016-08-13',
+      url: 'https://aotu.io/index.html'
+    },
+    {
+      title: '前端笔记',
+      date: '2016-08-13',
+      url: 'http://www.w3cmark.com/'
+    },
+    {
+      title: '百度EFE',
+      date: '2016-08-13',
+      url: 'http://efe.baidu.com/'
+    },
+    {
+      title: '腾讯Alloy',
+      date: '2016-08-13',
+      url: 'http://alloyteam.github.io/'
+    },
+    {
+      title: '前端网',
+      date: '2016-08-13',
+      url: 'http://www.qdfuns.com/'
+    },
+    {
+      title: '前端里',
+      date: '2016-08-13',
+      url: 'http://www.yyyweb.com/front'
+    },
+    {
+      title: '张云龙博客',
+      date: '2016-08-13',
+      url: 'https://github.com/fouber/blog'
+    }
+  ]
+};